app/ui/requirements.txt

streamlit
requests
Pillow
numpy
importlib-metadata>=1.0; python_version < "3.8"
